Zip config files to sdist
Added the logic of wrapping up the secretfy config files along with source.
While installation these files will be placed at
/usr/local/lib/python3.7/site-packages/ and in case of mock files generation
these files will be referred.

Also in case of `secretfy --mock` those files will not be added to git exclude
since the installed binary is not a git repository.

A config file is created which will provide the absolute path of the installed
directory. This logic is only for --mock implementation.